+++ include/clixon_custom.h
@@ -201,3 +201,22 @@
  * Introduced in 6.1, remove in 6.2
  */
 #define AUTOCLI_DEPRECATED_HIDE
+
+/*! Unified netconf input function
+ * Replace external, internal and controller netconf eventually
+ * New file: clixon_netconf_input.c with functions:
+ * - netconf_input_read2
+ * - netconf_input_msg2
+ * - netconf_input_frame2
+ * The following code should use this:
+ * X - netconf_main.c external netconf
+ * - netconf_proto.c internal netconf
+ * - controller device control
+ */
+/*! Use unified netconf input function for external use
+ */
+#define NETCONF_INPUT_UNIFIED_EXTERN
+
+/*! Use unified netconf input function for internal use
+ */
+#undef NETCONF_INPUT_UNIFIED_INTERNAL